Clean: clean and move some funcs to new files
