+++ src/push_swap.c
@@ -6,36 +6,11 @@
 /*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/05 13:12:05 by mbartos           #+#    #+#             */
-/*   Updated: 2024/01/12 16:19:34 by mbartos          ###   ########.fr       */
+/*   Updated: 2024/01/12 16:26:35 by mbartos          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
-
-// void	put_stck(t_node *stck)
-// {
-// 	ft_printf("-----\n");
-// 	while (stck != NULL)
-// 	{
-// 		ft_putnbr_fd(stck->number, 1);
-// 		ft_putchar_fd('\n', 1);
-// 		stck = stck->next;
-// 	}
-// 	ft_printf("-----\n");
-// }
-
-int	arr_length(char	**arr)
-{
-	int	i;
-
-	i = 0;
-	while (*arr != NULL)
-	{
-		i++;
-		arr++;
-	}
-	return (i);
-}
 
 int	ft_pushswap_atoi(const char *str)
 {
@@ -64,19 +39,6 @@ int	ft_pushswap_atoi(const char *str)
 		return (number);
 	ft_putstr_fd("Error: at least one number is not integer", 1);
 	exit(1);
-}
-
-void	free_array(char **arr)
-{
-	int	i;
-
-	i = 0;
-	while (arr[i] != NULL)
-	{
-		free(arr[i]);
-		i++;
-	}
-	free(arr);
 }
 
 t_node	*load_in_list(int argc, char **argv)
@@ -113,33 +75,6 @@ t_node	*load_in_list(int argc, char **argv)
 	}
 	return (stck);
 }
-
-// void	put_both_stck(t_node *stck_a, t_node *stck_b)
-// {
-// 	static int	fd;
-
-// 	if (fd == 0)
-// 	{
-// 		fd = open("output.txt", O_WRONLY | O_CREAT | O_TRUNC, 0666);
-// 		fd = -1; // delete if you want to write stacks to output.txt
-// 	}
-// 	ft_putstr_fd("-----\n", fd);
-// 	ft_putstr_fd("--A--\n", fd);
-// 	while (stck_a != NULL)
-// 	{
-// 		ft_putnbr_fd(stck_a->number, fd);
-// 		ft_putstr_fd("\n", fd);
-// 		stck_a = stck_a->next;
-// 	}
-// 	ft_putstr_fd("--B--\n", fd);
-// 	while (stck_b != NULL)
-// 	{
-// 		ft_putnbr_fd(stck_b->number, fd);
-// 		ft_putstr_fd("\n", fd);
-// 		stck_b = stck_b->next;
-// 	}
-// 	ft_putstr_fd("-----\n\n", fd);
-// }
 
 int	main(int argc, char **argv)
 {
