+++ src/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/21 12:27:47 by mbartos           #+#    #+#             */
-/*   Updated: 2024/01/12 16:01:35 by mbartos          ###   ########.fr       */
+/*   Updated: 2024/01/12 16:27:22 by mbartos          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,20 +44,25 @@ typedef struct node
 // init.c
 void	init_possblts(t_poss *possibilities);
 
+// array_utils.c
+int		arr_length(char	**arr);
+void	free_array(char **arr);
+
 // errors.c
 void	check_dup_nums(t_node *stck);
 void	check_args(int argc, char **argv);
 
 // ft_stck.c
-int		ft_stcksize(t_node *lst);
 t_node	*ft_stcknew(int number);
 void	ft_stckadd_back(t_node **lst, t_node *new);
-t_node	*ft_stcklast(t_node *lst);
-t_node	*ft_stck_seclast(t_node *lst);
-void	ft_freestck(t_node *lst);
+void	ft_stckclear(t_node **lst);
+
+// ft_stck_utils.c
+int		ft_stcksize(t_node *lst);
 int		ft_stck_min(t_node *lst);
 int		ft_stck_max(t_node *lst);
-void	ft_stckclear(t_node **lst);
+t_node	*ft_stcklast(t_node *lst);
+t_node	*ft_stck_seclast(t_node *lst);
 
 // pa_pb.c
 void	pa(t_node **stck_a, t_node **stck_b);
